Add the layer definitions when creating the metarig.
Also add support for Reload Scripts.
